Nearest donation/bank distance fix.
Think it works now...

+++ API/API/Models/Donation.cs
@@ -55,7 +55,7 @@ namespace API.Models
     {
         [Required]
         [Display(Name = "Distance")]
-        public float Distance { get; set; }
+        public Double Distance { get; set; }
     }
 
     public enum StatusType {
